pyproject.toml
[tool.poetry]
name = "smbclient"
version = "0.1.3"
description = ""
authors = ["d.a.kolesov <d.a.kolesov@s7.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pysmb = "^1.2.9.1"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
